app-movil/backend-cnn/requirements.txt
Flask==2.2.3
tensorflow==2.13.0
Werkzeug==2.2.3
requests==2.28.2
opencv-python==4.11.0.86
flask-cors
Pillow
pydicom
numpy